new module that enables debugging for every new ua
The idea behind this is that it can be used in any program. Right now it's not possible to configure it, but I am planning to add some kind of store for created user agents, so that it's possible to retrieve them via some imported DSL. That way it would be possible to change settings for individual user agents, which might be especially useful if there are several of those in a large application.