Re-log the request headers when we get the response
Because of the way the handlers work, we may not have logged all of the
request headers when we logged the request.  However, we'll keep sending
two version of the headers to guard against the case where a response
times out.  In that case we'd at least want to see *some* header
information about the request, even if it is limited.